Move servo to separate folder
