+++ src/main.cpp
@@ -14,6 +14,7 @@
 #include "menu/menu.h"
 #include "menu/def.h"
 #include "math/math_extensions.h"
+#include "servo/servo.h"
 #include "HX711.h"
 #include <EncButton.h>
 #include <avr/io.h>
@@ -144,13 +145,6 @@ enum stateS
     STORAGE,
     AUTOPID,
     NTC_ERROR,
-};
-
-enum stateServo
-{
-    CLOSED,
-    MOVE,
-    OPEN,
 };
 
 #define MAX_ERROR 30
@@ -409,153 +403,6 @@ public:
     }
 
 } iDryer;
-
-class servo
-{
-public:
-    uint8_t pin = 0;
-    unsigned long servoOldTime = 0;
-
-    uint16_t openTime = 0;
-    uint16_t closedTime = 0;
-
-    uint8_t angleMultiplier = 1;
-    uint16_t closedAngle = 90 * angleMultiplier;
-    uint16_t angle = 90;
-    uint16_t currentAngle = 90;
-
-    uint8_t changeState = 0;
-    uint8_t servoPin = 0;
-    uint16_t pulseWidth = 0;
-    stateServo prevState = OPEN;
-    stateServo state = CLOSED;
-
-    servo(uint8_t _srvPin, uint16_t _closedTime, uint16_t _openTime, uint16_t _angle)
-    {
-        pin = _srvPin;
-        closedTime = _closedTime;
-        openTime = _openTime;
-        angle = _angle * angleMultiplier;
-    }
-
-    void set(uint16_t _closedTime, uint16_t _openTime, uint16_t _angle)
-    {
-        closedTime = _closedTime;
-        openTime = _openTime;
-        angle = _angle * angleMultiplier;
-    }
-
-    void updateServo()
-    {
-        // if (setAngle(currentAngle, angle))
-        if (setAngle())
-        {
-            if (changeState)
-            {
-
-                if (currentAngle == closedAngle)
-                {
-                    state = CLOSED;
-                }
-                else
-                {
-                    state = OPEN;
-                }
-                changeState = 0;
-            }
-        }
-        else
-        {
-            if (servoPin)
-            {
-                PORTD &= ~(1 << pin);
-                Timer1.setPeriod(SERVO_PERIOD_MS * 1000 - pulseWidth);
-                servoPin = LOW;
-            }
-            else
-            {
-                pulseWidth = map(currentAngle, 0, 180 * angleMultiplier, SERVO_MIN_PULSE, SERVO_MAX_PULSE);
-                PORTD |= (1 << pin);
-                Timer1.setPeriod(pulseWidth);
-                servoPin = HIGH;
-            }
-        }
-    }
-
-    // bool setAngle(uint16_t &currentAngle, uint16_t &angle)
-    bool setAngle()
-    {
-        if (currentAngle != angle)
-        {
-            if (state != MOVE)
-            {
-                prevState = state;
-                state = MOVE;
-                PORTD &= ~(1 << pin);
-                Timer1.enableISR(CHANNEL_A);
-            }
-            if (currentAngle < angle)
-            {
-                currentAngle++;
-            }
-            else
-            {
-                currentAngle--;
-            }
-            return 0;
-        }
-        else
-        {
-            return 1;
-        }
-    }
-
-    void close()
-    {
-        state = OPEN;
-        currentAngle = 130;
-        changeState = 1;
-        angle = closedAngle;
-        updateServo();
-    }
-
-    void toggle()
-    {
-        if (state == CLOSED)
-        {
-            changeState = 1;
-            angle = closedAngle + angle;
-        }
-        if (state == OPEN)
-        {
-            changeState = 1;
-            angle = closedAngle;
-        }
-        updateServo();
-    }
-
-    void check()
-    {
-        if (state == CLOSED)
-        {
-            if (servoOldTime < millis() && openTime)
-            {
-                toggle();
-                servoOldTime = millis() + (unsigned long)openTime * 1000UL * 60UL;
-            }
-            updateServo();
-        }
-        if (state == OPEN)
-        {
-            if (servoOldTime < millis() && closedTime)
-            {
-                toggle();
-                servoOldTime = millis() + (unsigned long)closedTime * 1000UL * 60UL;
-            }
-            updateServo();
-        }
-    }
-};
 
 servo Servo(SERVO_1_PIN, eeprom_read_word(&menuVal[DEF_SERVO_CLOSED]), eeprom_read_word(&menuVal[DEF_SERVO_OPEN]), eeprom_read_word(&menuVal[DEF_SERVO_CORNER]));
